+++ src/pages/login.tsx
@@ -1,0 +1,5 @@
+function Login() {
+  return <div></div>;
+}
+
+export default Login;
