+++ src/pages/signup.tsx
@@ -1,0 +1,5 @@
+function Signup() {
+  return <div></div>;
+}
+
+export default Signup;
